requirements.txt
certifi==2024.2.2
python-amazon-paapi==5.0.1
python-dateutil==2.8.2
six==1.16.0
urllib3==2.2.0
